+++ cli/pcluster/config_sanity.py
@@ -22,7 +22,7 @@ import json
 from botocore.exceptions import ClientError
 
 def get_partition(region):
-    if region == 'us-gov-west-1':
+    if region.startswith('us-gov'):
         return 'aws-us-gov'
     return 'aws'
 
@@ -183,7 +183,7 @@ def check_resource(region, aws_access_key_id, aws_secret_access_key, resource_ty
     # Batch Parameters
     elif resource_type == 'AWSBatch_Parameters':
         # Check region
-        if region in ['us-gov-west-1', 'eu-west-3', 'ap-northeast-3']:
+        if region in ['us-gov-west-1', 'us-gov-east-1', 'eu-west-3', 'ap-northeast-3']:
             print ("ERROR: %s region is not supported with awsbatch" % region)
             sys.exit(1)